Move CommentUIComposer and CommentViewAdapter to their own files
