+++ EssentialFeed/EssentialFeediOS/Comment UI/Controllers/CommentUIComposer.swift
@@ -1,0 +1,31 @@
+//
+//  CommentUIComposer.swift
+//  EssentialFeediOS
+//
+//  Created by Khoi Nguyen on 5/1/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import Foundation
+import UIKit
+import EssentialFeed
+
+public final class CommentUIComposer {
+	private init() {}
+	
+	public static func commentComposeWith(loader: CommentLoader) -> CommentViewController {
+		let presentationAdapter = CommentLoaderPresentationAdapter(commentLoader: loader)
+		 
+		let bundle = Bundle(for: CommentViewController.self)
+		let storyBoard = UIStoryboard(name: "Comment", bundle: bundle)
+		let commentViewController = storyBoard.instantiateInitialViewController() as! CommentViewController
+		commentViewController.delegate = presentationAdapter
+		let presenter = CommentPresenter(
+			loadingView: WeakRefVirtualProxy(commentViewController),
+			errorView: WeakRefVirtualProxy(commentViewController),
+			commentView: CommentViewAdapter(controller: commentViewController))
+		presentationAdapter.presenter = presenter
+		
+		return commentViewController
+	}
+}
